// src/app/models/usuario.ts
export class Usuario {
    codigo?: number;
    id?: string;
    nombre?: string;
    apellido?: string;
    telefono?: string;
    mail?: string;
    cedula?: string ;
    permiso?: number;
    clave?: string ;
    estado?: number;
    permisoCodigo?: number;
    empresa?: number;    
    permisos?: any;
    menus?: any;
    empresas?: any;
    grupo?: any;
}